FSM Vhdl Project
VHDL Description of a Finite State Mealy Machine created using the Quartus II Software. A Machine which can be used as a combinational Lock system or a Pattern Recognition. 
Ica.vhd is the VHDL File .. it can also be added to your project
Ica.vwf is the waveform file used for simulation and testing,
Ica.qpf is the quartus II project file